feat(#4): criação do layout inicial da aplicação (header)
