+++ frontend/public/index.html
@@ -24,6 +24,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
+      <link rel="stylesheet" href="../src/style/resetCss.css">
     <title>React App</title>
   </head>
   <body>
